+++ check_netscaler.pl
@@ -6,9 +6,9 @@
 #
 # https://github.com/slauger/check_netscaler
 #
-# Version: v1.5.1 (2018-06-10)
+# Version: v1.6.0 (2019-10-04)
 #
-# Copyright 2015-2018 Simon Lauger
+# Copyright 2015-2019 Simon Lauger
 #
 # Contributor:
 # bb-ricardo (github.com/bb-ricardo)
@@ -40,7 +40,7 @@ use Time::Piece;
 my $plugin = Monitoring::Plugin->new(
   plugin    => 'check_netscaler',
   shortname => 'NetScaler',
-  version   => 'v1.5.1',
+  version   => 'v1.6.0',
   url       => 'https://github.com/slauger/check_netscaler',
   blurb     => 'Nagios Plugin for Citrix NetScaler Appliance (VPX/MPX/SDX/CPX)',
   usage     => 'Usage: %s