+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## current branch only
 - return OK in state check if no vservers are configured (#44)
+- replaced depreacted calls from Perl-Nagios-Plugin with new names from Perl-Monitoring-Plugin
 
 ## v1.5.1 (2018-06-10)
 - fixed minor typo, made matches/matches_not support matching in arrays